+++ src/Phobos.Ext.cpp
@@ -269,13 +269,12 @@ DEFINE_HOOK(0x67D04E, GameSave_SavegameInformation, 0x7)
 	return 0;
 }
 
-DEFINE_HOOK(0x559F29, LoadOptionsClass_GetFileInfo, 0x8)
+DEFINE_HOOK_AGAIN(0x67FD9D, LoadOptionsClass_GetFileInfo, 0x7)
+DEFINE_HOOK(0x67FDB1, LoadOptionsClass_GetFileInfo, 0x7)
 {
-	if (!R->BL()) return 0x55A03D; // vanilla overridden check
-
-	REF_STACK(SavegameInformation, Info, STACK_OFFSET(0x400, -0x3F4));
-	Info.InternalVersion = Info.InternalVersion - SAVEGAME_ID;
-	return 0x559F29 + 0x8;
+	GET(SavegameInformation*, Info, ESI);
+	Info->InternalVersion = Info->InternalVersion - SAVEGAME_ID;
+	return 0;
 }
 
 #ifdef DEBUG